Issue #1: Ensure put operation always returns HTTP status.
Previously the HTTP status was not being returned in the manner expected
by the edit script. HTTP 200 (successful) resulted in return code 0
while any error code resulted in an exception and return code 1.

Instead the script now always returns the HTTP status code. A non-200
HTTP status code is not treated as an exception but rather returned just
like a successful code. The assumption is that the calling script(s)
will act accordingly based on the response code and that this logic is
not in the domain of the put script.