Issue #5: Added rudimentary support for retrieving a revision.
The topic class has been expanded to support retrieval of a specific
revision. All get_* methods now support an optional revision parameter.

If no revision is specified then the latest version of the topic is
retrieved. If a revision is specified then the topic as it existed at
that time is retrieved. If a revision is specified but the revision does
not exist than currently the scripts will return "None".