Issue #8: Modified scripts to comply with PEP8 guidelines.
While for the most part the scripts already complied with PEP8 there
were quite a few instances of lines greater than 80 characters in length
or containing unnecessary whitespace.

I've updated all of the Python scripts to be in compliance with PEP8 and
will aim to keep it that way from this point onwards.