Issue #5: Improved speed of revision retrieval.
When retrieving a revision based on topic ID and revision ID the get_*
methods now specify the revision as part of the GET request. Previously
JSON of all revisions for the topic was retrieved and filtered which was
relatively slow.

These changes allow tools like pressgang-get to retrieve revisions very
quickly where the revision id is already known.