Issue #14: Report revision when saving via pressgang-edit.
When saving topics the pressgang-put script was returning the revision
information but this was not being displayed to the user by
pressgang-edit.

The pressgang-edit script has been updated to display the information
returned by pressgang-put, not just the HTTP status value, and the
pressgang-put script has been updated to improve the wording of the
message returned.